refactor: split agent run service
